chore(fe): move admin code editor folder
