+++ apps/frontend/app/admin/_components/code-editor/EditorResizablePanel.tsx
